Task 0. memset

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
  */
 void print_number(int n)
 {
-	unsigned int num;
+	long int num;
 
 	num = n;
